requirements.txt
python-telegram-bot==20.7
supabase==2.0.3
python-dotenv==1.0.0